pyproject.toml
[project]
name = "zuu"
#version = "6.0.0"
description = "Zack's Useful Utilities"
authors = [
    { name = "ZackaryW", email = "36378555+ZackaryW@users.noreply.github.com" }
]
dependencies = []
readme = "README.md"
requires-python = ">= 3.12"

[project.optional-dependencies]
app = [
    "requests>=2.32.3",
]
io = [
    "toml>=0.10.2",
    "pyyaml>=6.0.2"
]
date = [
    "croniter>=6.0.0",
    "dateparser>=1.2.0",
]
screen = [
    "pygetwindow>=0.0.9",
    "screeninfo>=0.8.1",
    "psutil>=6.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.4",
    "pillow>=11.1.0",
    "keyring>=25.6.0",
    "pywin32>=308",
    "pyscreeze>=1.0.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/zuu"]

[tool.ruff]
ignore = ["F405"]
